Conflicts:
	src/cpacs_other/CCPACSMaterial.cpp
	src/cpacs_other/CCPACSMaterial.h
	src/cpacs_other/TixiSaveExt.cpp
	src/cpacs_other/TixiSaveExt.h
	src/exports/CTiglExportIges.cpp
	src/geometry/CCPACSFarField.cpp
	src/geometry/CTiglAbstractGeometricComponent.cpp
	src/guide_curves/CCPACSGuideCurve.cpp
	src/guide_curves/CCPACSGuideCurve.h
	src/guide_curves/CCPACSGuideCurveProfile.cpp
	src/guide_curves/CCPACSGuideCurveProfile.h
	src/guide_curves/CCPACSGuideCurveProfiles.cpp
	src/guide_curves/CCPACSGuideCurves.cpp
	src/wing/CCPACSPositionings.h
	src/wing/CCPACSWing.cpp
	src/wing/CCPACSWingCSStructure.cpp
	src/wing/CCPACSWingCSStructure.h
	src/wing/CCPACSWingCell.cpp
	src/wing/CCPACSWingCell.h
	src/wing/CCPACSWingCells.cpp
	src/wing/CCPACSWingCells.h
	src/wing/CCPACSWingComponentSegment.cpp
	src/wing/CCPACSWingComponentSegment.h
	src/wing/CCPACSWingComponentSegments.cpp
	src/wing/CCPACSWingConnection.cpp
	src/wing/CCPACSWingConnection.h
	src/wing/CCPACSWingPositioning.cpp
	src/wing/CCPACSWingPositionings.cpp
	src/wing/CCPACSWingProfile.cpp
	src/wing/CCPACSWingProfile.h
	src/wing/CCPACSWingProfileCST.cpp
	src/wing/CCPACSWingProfileFactory.cpp
	src/wing/CCPACSWingProfileFactory.h
	src/wing/CCPACSWingProfilePointList.cpp
	src/wing/CCPACSWingProfiles.cpp
	src/wing/CCPACSWingSection.cpp
	src/wing/CCPACSWingSectionElement.cpp
	src/wing/CCPACSWingSectionElements.cpp
	src/wing/CCPACSWingSectionElements.h
	src/wing/CCPACSWingSections.cpp
	src/wing/CCPACSWingSections.h
	src/wing/CCPACSWingSegment.cpp
	src/wing/CCPACSWingSegment.h
	src/wing/CCPACSWingSegments.cpp
	src/wing/CCPACSWingSegments.h
	src/wing/CCPACSWingShell.cpp
	src/wing/CCPACSWingShell.h
	src/wing/CCPACSWings.cpp

+++ src/math/tiglmathfunctions.cpp
@@ -192,7 +192,7 @@ double class_function_deriv(const double& N1, const double& N2, const int& n, co
 double shape_function(const std::vector<double>& B, const double& x)
 {
     double ret = 0.;
-    int order = B.size()-1;
+    int order = static_cast<int>(B.size()) - 1;
     int i = 0;
     for (std::vector<double>::const_iterator bIT = B.begin(); bIT != B.end(); ++bIT, ++i) {
         ret += *bIT * bernstein_poly(i, order, x);
@@ -206,7 +206,7 @@ double shape_function(const std::vector<double>& B, const double& x)
 double shape_function_deriv(const std::vector<double>& B, const int& n, const double& x)
 {
     double ret = 0.;
-    int order = B.size()-1;
+    int order = static_cast<int>(B.size()) - 1;
     int i = 0;
     for (std::vector<double>::const_iterator bIT = B.begin(); bIT != B.end(); ++bIT, ++i) {
         ret += *bIT * tigl::bernstein_poly_deriv(n, i, order, x);
